iterate over length sorted list
When matching for submodule repos, sort by length. This will ensure the
.endswith(str) will match toplevel submodules before nested submodules
of the same name.

Nested modules can still be matched, but the user will need to specify
that in markdown by supplying the full path, exp:

[!git!submodule-hash url=https://github.com/idaholab/moose](contribs/moose)

instead of

[!git!submodule-hash url=https://github.com/idaholab/moose](moose)

Of course, (moose) will still match properly, if this is the only 'moose'
submodule in the repo.

Closes #29097