Add ability to override Mesh Generators when FileMesh is present
When a `FileMesh` and Mesh Generators are both simultaneously present in an input, a warning is raised in the `setup_mesh` task of `SetupMeshActions` because these types are not meant to be used with each other. In the following `set_mesh_base` task of `SetupMeshActions`, the mesh base is either:
1. set to an existing base from the `mesh_generator_system`, if Mesh Generators are present, or
2. built on-the-fly if Mesh Generators are not present.

Approach (1) is currently used when both `FileMesh` and Mesh Generator objects are present. Because the aforementioned warning is phrased to imply that  FileMesh objects do not use Mesh Generators, and because `FileMesh` objects are the default mesh type absent user input, an argument can be made that approach (2) should be used instead when `FileMesh` and Mesh Generators are simultaneously present.

This PR makes the change from approach (1) to (2) and makes the wording in the warning more explicit to reflect this change.

Consider the following case that prompted this change. The input file contains a `FileMeshGenerator` (i.e., a Mesh Generator). For restart purposes, the user wants to change the mesh file to a checkpoint, doing so by specifying `Mesh/file=cp/LATEST` on the command line. This results in an input with both a `FileMesh` and a Mesh Generator (the `FileMeshGenerator`). MOOSE raises the warning:
> Mesh Generators present but the [Mesh] block is set to construct a "FileMesh" mesh, which does not use Mesh Generators in constructing the mesh.

This warning implies that the `FileMesh` will be constructed instead of the Mesh Generator (`FileMeshGenerator`). However, the `SetupMeshAction` then proceeds to set the mesh base according to option (1), which utilizes the mesh base from the `mesh_generator_system` instead of building the mesh base on-the-fly, which is the correct approach for a `FileMesh` object. Finally, in the `init_mesh` task of `SetupMeshAction`, the `init` method of the `FileMesh` attempts to read the mesh in from the checkpoint file. However, because the mesh base was not correctly set, it already contains a nonzero number of elements. libMesh’s `CheckpointIO::read` method doesn’t like that the mesh object it is to populate from the checkpoint file already contains elements, so it throws a nondescript assertion error that took yours truly a long time to get to the bottom of. This PR is my solution to the issue.

In the input file of the above example, the mesh is not modified with other generators after the `FileMeshGenerator`. In this case, the obvious solution is to simply change the `FileMeshGenerator` in the input to a `FileMesh` to completely avoid having Mesh Generators in the first place. However, there could be more complex use cases that benefit from this PR.

There is also the reverse situation: where a `FileMesh` is specified in the input file, a `FileMeshGenerator` is specified on the command line. In this case, the `FileMesh` will stiff have precedence, and the warning will state this. If the user wants their command line option to be effective, they should override the input file’s existing `FileMeshGenerator` instead of trying to use a `FileMesh`.

Closes #29992